FebusTools

Documentation for FebusTools.

FebusTools._A1_ELTYPE

FebusTools.read_hdf5

FebusTools._A1_ELTYPE — Type

Element type of Febus A1 data

FebusTools.read_hdf5 — Method

read_hdf5(file; xlim=(-Inf, Inf), xdecimate=1, tlim=(-146138511-01-01T00:00:00, 146138512-12-31T23:59:59), blocks, zones=(==(1)), sources=(==(1)), version=nothing) -> ::FebusData

Read strain or strain rate data from file on disk. By default, all data are read into memory.

If xlim is given a length-2 collection, then only channels whose distances lie between xlim[begin] and xlim[end] will be read. Likewise, only blocks with times between tlim[begin] and tlim[end] are read. Note that whole blocks (usually of duration 1 s) are always read, so the stand and end times of the returned data can be respectively before and after xlim[begin] and xlim[end]. To read only the nth channel, set xdecimate=n.

Instead of tlim, you can also choose which blocks are read in with the blocks keyword argument, which again takes two items, the start and end block index (which begins at 1). Block indices outside the range of the data are warned about and only existing data which match the range are read.

To limit data to a set of zones, supply a function of form f(::Int)::Bool to zones which returns true when its first argument is the number of the zone to accept.

Likewise, supply a function of the form f(::Int)::Bool for sources to accept source a set of source numbers.

The data type (i.e., strain or strain rate) is recorded in the :data_type field of the metada dictionary.

By default, the file version is determined from the HDF5 file attributes, noting that files without a version attribute are considered to be 'v1' files. Pass a version (e.g., v"2.1") to force the assumption of a specific file version when parsing the data.

Note

Currently only the first source of the first zone is read, and an error is thrown if more than one zone or source is present in a file or if anything other than the first zone is requested. This will change in a breaking update to FebusTools.