FebusTools

Documentation for FebusTools.

FebusTools.read_hdf5

FebusTools.read_hdf5 — Function

read_hdf5(file[, cut1[, cut2]]; zones, sources) -> ::FebusData

Read strain or strain rate data from file on disk, reading blocks between dates cut1 and cut2. By default, all data are read.

To limit data to a set of zones, supply a function of form f(::Int)::Bool to zones which returns true when its first argument is the number of the zone to accept.

Likewise, supply a function of the form f(::Int)::Bool for sources to accept source a set of source numbers.

Note

Currently only the first source of the first zone is read, and an error is thrown if more than one zone or source is present in a file.